Execute feature extraction using temporal covariate settings.

This function executes the feature extraction on one or more cohortId in a cohort table and returns covariate data.

executeFeatureExtraction(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortIds,
  cohortTable,
  covariateSettings = NULL,
  addCohortBasedTemporalCovariateSettings = TRUE,
  covariateCohortDatabaseSchema = cohortDatabaseSchema,
  includeCovariateIds = NULL,
  covariateCohortTable = cohortTable,
  covariateCohortDefinitionSet = NULL,
  cohortCovariateAnalysisId = 150,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  outputFolder,
  aggregated = TRUE,
  rowIdField = "subject_id",
  incremental = TRUE
)

Arguments

connectionDetails

An object of type connectionDetails as created using createConnectionDetails. Can be left NULL if connection is provided. Both cannot be NULL.

connection

An object of type connection as created using connect. Can be left NULL if connectionDetails is provided. Both cannot be NULL.

cdmDatabaseSchema

Schema name where your patient-level data in OMOP CDM format resides. For SQL Server, include both database and schema name (e.g., 'cdm_data.dbo').

vocabularyDatabaseSchema

Schema name where your OMOP vocabulary tables reside. For SQL Server, include both database and schema name (e.g., 'vocabulary.dbo').

cohortDatabaseSchema

Schema name where your cohort tables reside. For SQL Server, include both database and schema name (e.g., 'scratch.dbo').

cohortIds

Vector of cohort IDs for which covariate extraction is to be performed.

cohortTable

Name of the table containing cohort data.

covariateSettings

A FeatureExtraction covariateSettings object. If NULL, cohort-based temporal covariate settings will be used.

addCohortBasedTemporalCovariateSettings

Logical flag indicating whether to add cohort-based temporal covariate settings. Default is TRUE.

covariateCohortDatabaseSchema

Schema name where covariate cohort data resides. Default is the same as cohortDatabaseSchema.

includeCovariateIds

Optional vector of covariate IDs to include.

covariateCohortTable

Name of the table containing covariate cohort data. Default is the same as cohortTable.

covariateCohortDefinitionSet

A data frame defining covariate cohort definitions used to generate covariate IDs.

cohortCovariateAnalysisId

An integer identifier used to generate covariate IDs. Default is 150.

tempEmulationSchema

Some database platforms (e.g., Oracle, Impala) do not support temporary tables. Provide a schema with write privileges where temporary tables can be created. Default is obtained from getOption("sqlRenderTempEmulationSchema").

outputFolder

Name of the local folder to place results. Use forward slashes (/); avoid network drives.

aggregated

Logical flag indicating whether to aggregate covariate data. Default is TRUE.

rowIdField

Field name used as the row identifier (e.g., "subject_id" or "row_id"). Default is "subject_id".

incremental

Logical flag indicating whether to skip processing if output for a cohort already exists. Default is TRUE.

Value

No explicit return value. The function saves the covariate data files to the specified output folder.

Details

The function performs the following tasks:

Establishes a connection to the database using connection details if no connection is provided.

Validates and creates the output folder if it does not exist.

Checks that only one cohort is provided if aggregated data is disabled.

Configures cohort-based temporal covariate settings when requested.

Optionally creates a temporary table (when using a row identifier other than "subject_id") to store cohort data with row numbers.

Iterates over each cohort ID and extracts covariate data via FeatureExtraction::getDbCovariateData.

Saves the extracted covariate data to a zip file in the output folder.

Cleans up any temporary tables that were created.

Examples

if (FALSE) { # \dontrun{
executeFeatureExtraction(
  connectionDetails = myConnectionDetails,
  cdmDatabaseSchema = "cdm_data.dbo",
  cohortDatabaseSchema = "scratch.dbo",
  cohortIds = c(1, 2, 3),
  cohortTable = "cohort_table",
  outputFolder = "results/output",
  covariateSettings = myCovariateSettings
)
} # }